+++ source/Erwin/core/layer_stack.cpp
@@ -21,7 +21,7 @@ void LayerStack::clear()
 
 size_t LayerStack::push_layer(Layer* layer)
 {
-	layers_.emplace(layers_.begin() + overlay_pos_, layer);
+	layers_.emplace(layers_.begin() + long(overlay_pos_), layer);
 	++overlay_pos_;
 	layer->on_attach();
 
@@ -56,7 +56,7 @@ void LayerStack::pop_layer(size_t index)
 		DLOG("application",1) << "Popped layer \"" << WCC('n') << layer->get_name() << WCC(0) << "\" at index " << index << std::endl;
 		
 		delete layer;
-		layers_.erase(layers_.begin() + index);
+		layers_.erase(layers_.begin() + long(index));
 		--overlay_pos_;
 
 		update_layer_ids();
@@ -74,7 +74,7 @@ void LayerStack::pop_overlay(size_t index)
 		DLOG("application",1) << "Popped overlay \"" << WCC('n') << layer->get_name() << WCC(0) << "\" at index " << index << std::endl;
 
 		delete layer;
-		layers_.erase(layers_.begin() + index);
+		layers_.erase(layers_.begin() + long(index));
 
 		update_layer_ids();
 	}
@@ -96,10 +96,10 @@ void LayerStack::set_layer_enabled(size_t index, bool value)
 std::ostream& operator <<(std::ostream& stream, const LayerStack& rhs)
 {
 	stream << "[ ";
-	for(auto it = rhs.layers_.begin(); it<rhs.layers_.begin()+rhs.overlay_pos_; ++it)
+	for(auto it = rhs.layers_.begin(); it<rhs.layers_.begin()+long(rhs.overlay_pos_); ++it)
 		stream << "{" << (*it)->get_name() << "} ";
 	stream << "| ";
-	for(auto it = rhs.layers_.begin()+rhs.overlay_pos_; it<rhs.layers_.end(); ++it)
+	for(auto it = rhs.layers_.begin()+long(rhs.overlay_pos_); it<rhs.layers_.end(); ++it)
 		stream << "{" << (*it)->get_name() << "} ";
 	stream << "]";
